authentication/requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
fastapi==0.110.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
multidict==6.0.5
psycopg==3.1.18
psycopg-binary==3.1.18
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
yarl==1.9.4
